Windows: utilize ShellExecuteW
On Windows, use the `ShellExecuteW` syscall instead of an external command. The
`HideWindow` option is no longer configurable, as `setFlags` assumes an external
command is to be run. Long term, would be better to refactor the package, as
different Window behavior can be configured via the last argument to
`ShellExecuteW` (`nShowCmd`).

https://docs.microsoft.com/windows/win32/api/shellapi/nf-shellapi-shellexecutew

Fixes #16